Add view to edit DeliveryNote
