+++ lupanes/templates/lupanes/deliverynote_form.html
@@ -15,6 +15,7 @@
       <th>#</th>
       <th>Producto</th>
       <th>Cantidad</th>
+      <th></th>
     </tr>
   </thead>
   <tbody>
@@ -23,6 +24,7 @@
       <td>{{ forloop.counter }}</td>
       <td>{{ note.product.name }}</td>
       <td>{{ note.quantity }} {{ note.product.unit }}</td>
+      <td><a class="btn btn-outline-warning" href="{% url 'lupanes:deliverynote-edit' note.pk %}">editar</a></td>
     </tr>
   {% endfor %}
   </tbody>
